fix: harden GPU-targeted plans and runtime state
Compile execution plans against the selected context and include semantic graph, build options, runtime mode, and cooperative-matrix capabilities in cache invalidation.

Validate checkpoint storage before mapped copies, key attention pipelines by head width, synchronize host uploads, release all owned resources, and replace invalid device-scope gradient-clip atomics with barriered f32 accumulation.

Add mixed-width attention, malformed-checkpoint, f16-only target, SPIR-V, and AMD portability regressions while bringing existing Rust sources to the enforced format/lint baseline.

+++ src/shaders/grad_clip_zero.wgsl
@@ -2,12 +2,11 @@
 // gradient-norm-squared accumulator for grad clipping.
 // Run once per step() before any GradClipNormSq dispatches.
 //
-// `acc[0]` holds f32 sum-of-squares stored as u32 bits; writing 0u is
-// equivalent to writing 0.0f (positive-zero).
+// `acc[0]` holds the f32 sum-of-squares.
 
-var<storage, read_write> acc: array<u32>;
+var<storage, read_write> acc: array<f32>;
 
 @compute @workgroup_size(1)
 fn main() {
-    acc[0] = 0u;
+    acc[0] = 0.0;
 }
